Refactor other functions
